task

+++ utils/redis.js
@@ -3,19 +3,37 @@ import { createClient } from 'redis';
 class RedisClient {
   constructor(){
     this.client = createClient();
+    this.connected = false;
+
+    try {
+      this.connect();
+    } catch (e) {
+      console.error(`There is Error in connection: ${e}`);
+    }
   }
 
+  connect() {
+    return new Promise((resolve, reject) => {
+      (async () => {
+        try {
+          this.client.connect();
+          this.connected = true;
+          resolve();
+        } catch (e) {
+          reject('Failed to connect to the server');
+        }
+      })();
+    })
+  }
   isAlive() {
-    this.client.connect()
-    if (this.client.isOpen) {
-      return true;
-    } else {
-      return false;
-    }
+    return this.connected && this.client.isOpen;
   }
 
   async get(key) {
     try {
+      if (!this.isAlive()) {
+        return null;
+      }
       return await this.client.get(key);
     } catch (e) {
       console.error(`Get Error:${e}`);
@@ -24,6 +42,9 @@ class RedisClient {
 
   async set(key, value, duration) {
     try {
+      if (!this.isAlive()) {
+        return null;
+      }
       await this.client.set(key, value, {
         EX: duration,
       });
@@ -34,11 +55,14 @@ class RedisClient {
 
   async del(key) {
     try {
+      if (!this.isAlive()) {
+        return null;
+      }
       await this.client.del(key);
     } catch (e) {
       console.error(`Del Error:${e}`);
     }
   }
 }
-const redisClient = new RedisClient;
+const redisClient = new RedisClient();
 export default redisClient;
